Docs: rename `webdocs` package to `website`
